Fix AMF encoder to work properly with pointer/non-pointer/nil types
There were two issues identified here. The first was that rather than trying
to set the value of the `amft` pointer, the marshaller was trying to set the
pointer itself, which is unaddressable. The second was that the identifier
was unreferencing pointer types when trying to identify them, rather than
referencing non-pointer types (everything in the typs map is a pointer). It
was necessary to add a new method, NewMatchingTypeFromValue, to maintain
addressability; values returned from value.Interface() are not addressable.